backend/requirements.txt
# Backend requirements for Background Worker

# Google Gemini API
google-generativeai

# Database
supabase

# Sentiment Analysis
transformers>=4.35.0
torch>=2.1.0
scipy>=1.11.0
numpy>=1.24.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.3
